Create a border
The changes are made in order to specify where the location where game is taking place.

To do so, a Turtle object was created and set in location (-300,-300). It then draws a long line before turning counterclockwise, and continues to do so for a total of 4 times. The effect is that a white boundary was created, where all gameplay will occur within.

A limitation of this build is that Pac-Man still needs to be created