Added new feature
In this build, a feature was introduced that allows players to use arrow keys to control Pac-Man. First, four methods are created that allows for Pac-Man to change its x-y coordinates. Then, an onkey() method from turtle class is used to bind each directional key on keyboard to a method. Finally, listen() method was used to collect the onkey() events

An important limitation of this build is there are no enemies present