Modify protagonist object
When running the previous build, when its position has changed, the protagonist Pac-Man leaves a yellow line between origin and new location. This is problematic because it leaves a bad feature in the game. It was solved by moving the setposition() method below penup() method, which allows penup() method to erase the lines created by setposition() method.

In addition, when running previous build, Pac-Man initially appears in origin, before slowly moving to its initial set position. It is solved by using a speed() and set it to 0, which was the fastest speed. This allows Pac-Man to appear at initial set position instantaneously

A limitation is that there are still no enemies present.