Creates an enemy object
In this build, an enemyone object was created, which has the form of a red triangle facing downwards. At the moment, a triangle was used to represent enemy, which will be modified in the future. This creates an enemy avatar that could potentially serve as an obstacle

A limitation of this build is the enemy cannot move.