Add a new feature
In the previous build, the enemy is not moving, and therefore is not seen as a challenge to player. In this build, several if statements are used to compare enemy's x and y location with pac-man's x and y location. Compared to pac-man's x coordinate, the enemy will increase its x-coordinate if smaller, or decrease if bigger. Same thing is done with y-coordinate.

This leads to enemy adjusting its horizontal and vertical position until it reaches the position of player's pac-man.

In addition, a random variable is assigned a value of 0 or 1. If value is 0, then enemy will first move left or right until it has same x coordinate as pac-man, before moving up and down; if value is 1, then enemy will move up and down first until it has same y coordinate as pac-man, before moving left and right.

A possible limitation is that nothing happens when enemy reaches pac-man.