Add new feature
A problem with previous build is that nothing happens if the enemy avatar reaches pac-man.

To solve this problem, an if statement is used. If both pac-man and enemy have same x-coordinate and y-coordinate, an object will be created. The object will use its write method to write a message at center of screen.

As a result, if enemy reaches pac-man, a "Game Over" message will be presented at center of screen.

A possible limitation is that there are only one enemy present. In order to make games challenging, there needs to be more than one obstacle.