Add new feature
A problem with previous build is that there is only one enemy present.

To solve this problem, three more enemy objects are created. They have similar attribute compared to each other, but have different starting positions and color

As a result, there are now four enemy ghosts present on the screen.

A possible limitation is that only one enemy is actually chasing pac-man, while the other three enemies is stationary.