Fix a bug
In previous build, even after game over pac-man can still respond to player input and move.

In this build, in the if loop where enemy coordinates and pac-man coordinates are the same, onkey() methods are used again. This time, the four arrow keys are assigned no functions.

This means when game over screen appears, player will no longer be able to control pac-man.

An important limitation to this build is that not all four enemies are functional. Only one enemy actually chases pac-man.