Add new background features
A problem with previous build is that there are no walls within the game. To solve this problem, a Turtle object was created. The turtle starts at position (-270, -270) and draws a 60 by 60 square, with the starting point being square's bottom-left vertex. After drawing the square, the turtle then moves right by 100 before drawing another 60 by 60 square. This process continues until turtle's location exceeds 280.

The turtle will repeat the above step, but each time at a x-coordinate 100 units higher than before. The result of this is 36 large squares will be created, arranged in a 6 by 6 manner. These squares will serve as in game obstacles. In addition, a 30 to 40 unit gap will be present between these wall, which will allow Pac-Man and ghost to go through.

A limitation of this build is that despite intended to be a barrier, both Pac-Man and ghosts can walk through the wall.